app/requirements.txt
fastapi
uvicorn
PyJWT
sqlite-utils
structlog
pytest
httpx
